SDL 1.3 build fixes. Support SDL_CONFIG environment variable in update-workspaces to pick non-default sdl-config. Avoid non-standard use of SDL/ in include paths.
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@10895 3db68df2-c116-0410-a063-a993310a9797

+++ source/ps/CConsole.cpp
@@ -652,7 +652,7 @@ static bool isUnprintableChar(SDL_keysym key)
 
 InReaction conInputHandler(const SDL_Event_* ev)
 {
-	if (ev->ev.type == SDL_HOTKEYDOWN)
+	if ((int)ev->ev.type == SDL_HOTKEYDOWN)
 	{
 		std::string hotkey = static_cast<const char*>(ev->ev.user.data1);
